# api-service/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        Stock:
            properties:
                name:
                    type: string
                symbol:
                    type: string
                open:
                    type: number
                    format: double
                high:
                    type: number
                    format: double
                low:
                    type: number
                    format: double
                close:
                    type: number
                    format: double
            required:
                - name
                - symbol
                - open
                - high
                - low
                - close
            type: object
            additionalProperties: false
        ErrorResult:
            properties:
                message:
                    type: string
            required:
                - message
            type: object
            additionalProperties: false
        LoginResponse:
            properties:
                token:
                    type: string
            required:
                - token
            type: object
            additionalProperties: false
        RegisterUserResponse:
            properties:
                token:
                    type: string
                email:
                    type: string
            required:
                - token
                - email
            type: object
            additionalProperties: false
        RoleEnum:
            enum:
                - user
                - admin
            type: string
        Stats:
            properties:
                stock:
                    type: string
                timesRequested:
                    type: number
                    format: double
            required:
                - stock
                - timesRequested
            type: object
            additionalProperties: false
    securitySchemes: {}
info:
    title: api-service
    version: 1.0.0
    contact: {}
openapi: 3.0.0
paths:
    /v1/history:
        get:
            operationId: GetHistorical
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {items: {$ref: '#/components/schemas/Stock'}, type: array}
                                    - {$ref: '#/components/schemas/ErrorResult'}
                '400':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '404':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '500':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
            description: 'Retrivies an array of queries history.'
            tags:
                - History
            security: []
            parameters:
                -
                    in: header
                    name: authorization
                    required: true
                    schema:
                        type: string
    /v1/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/LoginResponse'}
                                    - {$ref: '#/components/schemas/ErrorResult'}
                '400':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '404':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '500':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
            description: 'Logins and retrieves an authorization token.'
            tags:
                - Login
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                                email:
                                    type: string
                            required:
                                - password
                                - email
                            type: object
    /v1/register:
        post:
            operationId: Register
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {$ref: '#/components/schemas/RegisterUserResponse'}
                                    - {$ref: '#/components/schemas/ErrorResult'}
                '400':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '404':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '500':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
            description: 'Register an user.'
            tags:
                - Register
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                password:
                                    type: string
                                role:
                                    $ref: '#/components/schemas/RoleEnum'
                                email:
                                    type: string
                            required:
                                - password
                                - role
                                - email
                            type: object
    /v1/stats:
        get:
            operationId: GetStats
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {items: {$ref: '#/components/schemas/Stats'}, type: array}
                                    - {$ref: '#/components/schemas/ErrorResult'}
                '400':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '404':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '500':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
            description: 'Retrivies an array of the most searched stocks.'
            tags:
                - Stats
            security: []
            parameters:
                -
                    in: header
                    name: authorization
                    required: true
                    schema:
                        type: string
    /v1/stock:
        get:
            operationId: GetStock
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                anyOf:
                                    - {items: {$ref: '#/components/schemas/Stock'}, type: array}
                                    - {$ref: '#/components/schemas/ErrorResult'}
                '400':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '404':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
                '500':
                    description: ""
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResult'
            description: 'Retrivies an array of stock information.'
            tags:
                - Stocks
            security: []
            parameters:
                -
                    in: query
                    name: stockCode
                    required: true
                    schema:
                        type: string
                -
                    in: header
                    name: authorization
                    required: true
                    schema:
                        type: string
servers:
    -
        url: 'http://localhost:3001/api-service/api'